requirements.txt
nameparser>=1.1.3
fuzzywuzzy>=0.18.0
python-levenshtein>=0.12.2
spacy>=3.4.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0